+++ apps/src/aichat/views/modelCustomization/ExampleTopicsInputs.tsx
@@ -1,0 +1,49 @@
+import React from 'react';
+
+import {useAppDispatch} from '@cdo/apps/util/reduxHooks';
+import MultiItemInput from '@cdo/apps/templates/MultiItemInput';
+
+import {setModelCardProperty} from '../../redux/aichatRedux';
+
+const ExampleTopicsInputs: React.FunctionComponent<{
+  topics: string[];
+  readOnly: boolean;
+}> = ({topics, readOnly}) => {
+  const dispatch = useAppDispatch();
+
+  return (
+    <MultiItemInput
+      key="exampleTopics"
+      items={topics}
+      onAdd={() =>
+        dispatch(
+          setModelCardProperty({
+            property: 'exampleTopics',
+            value: [...topics].concat(''),
+          })
+        )
+      }
+      onRemove={() => {
+        dispatch(
+          setModelCardProperty({
+            property: 'exampleTopics',
+            value: [...topics].slice(0, -1),
+          })
+        );
+      }}
+      onChange={(index, value) => {
+        const updatedTopics = topics.slice();
+        updatedTopics[index] = value;
+        dispatch(
+          setModelCardProperty({
+            property: 'exampleTopics',
+            value: updatedTopics,
+          })
+        );
+      }}
+      readOnly={readOnly}
+    />
+  );
+};
+
+export default ExampleTopicsInputs;
