+++ apps/src/lab2/levelEditors/aichatSettings/edit-aichat-settings.module.scss
@@ -1,5 +1,4 @@
 @import 'color.scss';
-@import '@cdo/apps/mixins.scss';
 
 .fieldSection {
   display: flex;
@@ -34,33 +33,4 @@
 .textarea {
   width: 100%;
   box-sizing: border-box;
-}
-
-.multiItemContainer {
-  display: flex;
-  flex-direction: column;
-}
-
-.multiItemInput {
-  display: flex;
-  flex-wrap: wrap;
-
-  &Multiline {
-    flex-direction: column;
-    flex-wrap: nowrap;
-  }
-}
-
-.buttonsRow {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-}
-
-.plusMinusButton {
-  @include remove-button-styles;
-  border: 1px solid $neutral_dark;
-  border-radius: 5px;
-  padding: 5px;
-  margin-right: 10px;
 }
